requirements.txt
fastapi
uvicorn
python-multipart
langgraph
langchain
langchain_community
langchain-nomic
langchain-groq
chromadb
python-dotenv
pydantic
tiktoken
pypdf2
pdfplumber